Permanent storage of state.
Figuring out the path where we can store things is fun.

On unixy things the current code should be sufficient. No attempt to
do Windows for now. No support for iOS (unless the unixy thing just
works, that would be nice).

I have no idea how to do this right on Android so it's just brutally
hardcoded to whatever works on my phone. It doesn't look right to me,
but it's quite hard to figure out how to do this right since Android
insists on documenting everything in terms of "call this Java
function" which isn't very helpful.